requirements.txt
news-please
qdrant-client
pymongo
schedule
pydantic-settings